relocated alignments module up one level changed corresponding import moved template library to data folder
