+++ freecad/trails/resources/snippets/template/SketchTemplate.py
