+++ freecad/trails/resources/snippets/template/TemplateLibrary.py
