+++ freecad/trails/resources/snippets/template/ViewTemplateLibrary.py
