+++ freecad/trails/resources/snippets/template/__init__.py
